🔒 Warden: fix integer overflows in distance calculations
- Fixed unhandled integer overflow during distance calculation (e.g. `dx * dx` or `a.x - b.x` causing `attempt to subtract with overflow` and `attempt to multiply with overflow`) allowed panics and DoS if extreme coordinates were passed in `curvature.rs`, `bureaucratic_black_hole.rs`, `radioactive.rs`, `orbital_mirrors.rs`, `hum.rs`, `lighting.rs`, and `seismic.rs`.
- Cast coordinate values to `f32` *before* subtraction or multiplication to ensure floating-point math handles large inputs without panicking.

+++ src/layer1/hum.rs
@@ -100,7 +100,7 @@ pub fn update_hum_system(mut hum_map: ResMut<HumMap>, sources: Query<(&HumSource
 
         for dy in -radius_i..=radius_i {
             for dx in -radius_i..=radius_i {
-                let dist_sq = (dx * dx + dy * dy) as f32;
+                let dist_sq = dx as f32 * dx as f32 + dy as f32 * dy as f32;
                 if dist_sq <= radius_sq {
                     let dist = dist_sq.sqrt();
                     // Linear falloff: 1.0 at center, 0.0 at radius
